requirement.txt
#Django
Django==4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
requests==2.32.4

#image
pillow==11.1.0

#auth
